added new algo. into : algorithms/java/Maths dir..
For a given number N check if it is prime or not. A prime number is a number which is only divisible by 1 and itself.
 Example 1:

Input:
N = 5
Output:
1
Explanation:
5 has 2 factors 1 and 5 only.

PR Checklist:

 My submission is formatted according to the guidelines in the contributing guide
 My addition is on refer on the language README.md file
 My addition does not have a spelling problem
 My submission has a proper and user-friendly description of the algorithm
 My submission has the time complexity of the algorithm
 My submission has sample input-output of the program (NOT FOR PYTHON)
What kind of change does this PR introduce? (check at least one)

 Bugfix
 New algorithm
 Optimization in previous algorithms
 Code style update
 Refactor
 Documentation
 Other, please describe:
Briefly describe the changes in this PR
code added for longest palindromic substring